Replaced EmployeeModel's list with hibernate calls.

+++ src/main/java/com/step/controller/ListController.java
@@ -3,7 +3,9 @@ package com.step.controller;
 import com.step.model.EmployeeModel;
 
 import javax.servlet.ServletException;
-import javax.servlet.http.*;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
 // TODO: afisarea, adaugarea, editarea si stergerea
@@ -48,7 +50,7 @@ public class ListController extends HttpServlet {
 
         // 3. Informatia din employee model se seteaza pe obiectul request
 
-        int totalPages = model.getTotalPages();
+        long totalPages = model.getTotalPages();
 
         req.setAttribute("list",  model.getPage(page));
         req.setAttribute("activePage", page);
